Added focus controls to the camera and added a universal list of active planets

+++ Orbit Sim 2D/Assets/Scripts/TimeKeeper.cs
@@ -14,7 +14,7 @@ public class TimeKeeper : MonoBehaviour
         instance = this;
     }
     #endregion
-    public float time;
+    public double time;
     // Start is called before the first frame update
     void Start()
     {
